Clean up barchart data link supported tooltip implementation
